rebase

+++ implicitmodules/torch/StructuredFields/StructuredField_0.py
@@ -16,5 +16,5 @@ class StructuredField_0(SupportStructuredField):
     def __call__(self, points, k=0):
         ker_vec = gauss_kernel(rel_differences(points, self.support), k, self.__sigma)
         ker_vec = ker_vec.reshape((points.shape[0], self.support.shape[0]) + tuple(ker_vec.shape[1:]))
-        return torch.tensordot(torch.transpose(torch.tensordot(torch.eye(2), ker_vec, dims=0), 0, 2), self.moments,
+        return torch.tensordot(torch.transpose(torch.tensordot(torch.eye(2, dtype=points.dtype), ker_vec, dims=0), 0, 2), self.moments,
                                dims=([2, 3], [1, 0]))
